Refactoring bufferq code with the same interface.
