--- bufferq/util.py
@@ -1,9 +1,0 @@
-"""util.py.
-
-Utilities for the bufferq module.
-"""
-import time
-
-def diff_time() -> float:
-    """Return a time suitable for 'timeout' comparisons."""
-    return time.monotonic()
